- replaced miles with fmod - couple fixes - esc key brings up game menu instead popup stuff, so make sure to get a buttonbar mod if you need them

+++ src/engine/shared/library/sharedUtility/src/shared/NameGenerator.cpp
@@ -633,10 +633,10 @@ bool NameGenerator::verifyName(const Unicode::String &name) const
 			namelength = name.size() - pos;
 		}
 		
-		if (namelength < m_nameTemplates[nameindex].m_charsMin)
+		if (namelength < static_cast<unsigned int>(m_nameTemplates[nameindex].m_charsMin))
 			return false;
 		
-		if (namelength > m_nameTemplates[nameindex].m_charsMax)
+		if (namelength > static_cast<unsigned int>(m_nameTemplates[nameindex].m_charsMax))
 			return false;
 		
 		pos = newpos + 1;
